Moved files and adapted controller
