+++ src/main/java/at/fhtw/swen3/persistence/DALConnectionException.java
@@ -1,4 +1,4 @@
-package at.fhtw.swen3.services.exceptions;
+package at.fhtw.swen3.persistence;
 
 public class DALConnectionException extends DALException {
     public DALConnectionException(String errorMessage, String functionName, String className) {
