+++ src/main/java/at/fhtw/swen3/services/ParcelService.java
@@ -1,10 +1,6 @@
 package at.fhtw.swen3.services;
 
 import at.fhtw.swen3.persistence.entities.ParcelEntity;
-import at.fhtw.swen3.services.exceptions.BLDataNotFoundException;
-import at.fhtw.swen3.services.exceptions.BLValidationException;
-import io.swagger.v3.oas.annotations.servers.Server;
-import org.springframework.stereotype.Service;
 
 import java.util.Optional;
