+++ src/main/java/at/fhtw/swen3/services/WarehouseService.java
@@ -2,12 +2,6 @@ package at.fhtw.swen3.services;
 
 import at.fhtw.swen3.persistence.entities.HopEntity;
 import at.fhtw.swen3.persistence.entities.WarehouseEntity;
-import at.fhtw.swen3.services.dto.Hop;
-import at.fhtw.swen3.services.dto.Warehouse;
-import at.fhtw.swen3.services.exceptions.BLDataNotFoundException;
-import at.fhtw.swen3.services.exceptions.BLValidationException;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Service;
 
 import java.util.Optional;
