+++ src/main/java/at/fhtw/swen3/services/impl/ParcelServiceImpl.java
@@ -4,9 +4,9 @@ import at.fhtw.swen3.persistence.entities.ParcelEntity;
 import at.fhtw.swen3.persistence.repositories.ParcelRepository;
 import at.fhtw.swen3.services.ParcelService;
 import at.fhtw.swen3.services.dto.StateEnum;
-import at.fhtw.swen3.services.exceptions.BLDataNotFoundException;
-import at.fhtw.swen3.services.exceptions.BLValidationException;
-import at.fhtw.swen3.services.exceptions.DALDataNotFondException;
+import at.fhtw.swen3.services.BLDataNotFoundException;
+import at.fhtw.swen3.services.BLValidationException;
+import at.fhtw.swen3.persistence.DALDataNotFondException;
 import at.fhtw.swen3.services.validation.InputValidator;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
